% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\name{collapse_authors}
\alias{collapse_authors}
\title{Collapse Quarto authors}
\usage{
collapse_authors(
  config = quarto::quarto_inspect()$config,
  reverse_first = TRUE,
  orcid_style = c("none", "icon", "plain"),
  orcid_icon = fontawesome::fa(name = "orcid"),
  sep = ", ",
  sep2 = " and ",
  last = ", and "
)
}
\arguments{
\item{config}{Quarto configuration as returned by \code{\link[quarto:quarto_inspect]{quarto::quarto_inspect()$config}}.}

\item{reverse_first}{Whether or not to reverse the first and last name from the \emph{first} author, meaning \verb{Jane Doe} becomes \verb{Doe, Jane}.}

\item{orcid_style}{How to include the authors' \href{https://en.wikipedia.org/wiki/ORCID}{ORCID}s in the output. One of
\itemize{
\item \code{"none"} to omit ORCIDs in the output.
\item \code{"icon"} to include an ORCID icon (specified via \code{orcid_icon}) after the author name that links to the respective author's ORCID page. Suitable for HTML
output.
\item \code{"plain"} to simply include the ORCID as plain text in parentheses after the author name.
}}

\item{orcid_icon}{SVG icon to be used for \code{orcid_style = "icon"}.}

\item{sep}{Separator. A character string.}

\item{sep2}{Separator for the special case that \code{x} contains only two
elements. A character string.}

\item{last}{Last separator, if there is no truncation. E.g. use
\code{", and "} for the \href{https://en.wikipedia.org/wiki/Serial_comma}{serial comma}. A character string.}
}
\value{
A character scalar.
}
\description{
Collapses all author names from a Quarto project or document into a single string.
}
\examples{
if (quarto::is_using_quarto()) {
  quappo::collapse_authors()
}
}
\seealso{
Other metadata: 
\code{\link{profile}()},
\code{\link{profiles}()},
\code{\link{qmd_layout_classes}}
}
\concept{metadata}
