% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\name{cur_profiles}
\alias{cur_profiles}
\title{Get current Quarto profile(s)}
\usage{
cur_profiles()
}
\value{
\code{NULL} when Quarto is not rendering, otherwise a character vector of the active Quarto profiles.
}
\description{
Returns the currently active Quarto profile(s). Wrapper around \code{Sys.getenv("QUARTO_PROFILE")} that returns \code{NULL} when Quarto is not rendering.
}
\examples{
quappo::cur_profiles()
}
\seealso{
Other metadata: 
\code{\link{collapse_authors}()},
\code{\link{profiles}()},
\code{\link{qmd_layout_classes}}
}
\concept{metadata}
