% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\name{default_profiles}
\alias{default_profiles}
\title{Get default Quarto project profile(s)}
\usage{
default_profiles(input = ".")
}
\arguments{
\item{input}{Path to a \href{https://quarto.org/docs/projects/quarto-projects.html}{Quarto project} directory. A character scalar.}
}
\value{
\code{NULL} if no default Quarto project profiles are configured, otherwise a character vector.
}
\description{
Returns a Quarto project's \href{https://quarto.org/docs/projects/profiles.html#default-profile}{default profile(s)}.
}
\examples{
if (quarto::is_using_quarto()) {
  quappo::default_profiles()
}
}
\seealso{
Other metadata: 
\code{\link{collapse_authors}()},
\code{\link{cur_profiles}()},
\code{\link{profiles}()},
\code{\link{qmd_layout_classes}}
}
\concept{metadata}
