% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\name{profile}
\alias{profile}
\title{Get current Quarto profile}
\usage{
profile()
}
\value{
\code{NULL} when Quarto is not rendering, otherwise a character scalar.
}
\description{
Returns the currently active Quarto profile. Wrapper around \code{Sys.getenv("QUARTO_PROFILE")} that returns \code{NULL} when Quarto is not rendering.
}
\examples{
quappo::profile()
}
\seealso{
Other metadata: 
\code{\link{collapse_authors}()},
\code{\link{profiles}()},
\code{\link{qmd_layout_classes}}
}
\concept{metadata}
