% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\name{profiles}
\alias{profiles}
\title{List Quarto project profiles}
\usage{
profiles(input = ".")
}
\arguments{
\item{input}{Path to a \href{https://quarto.org/docs/projects/quarto-projects.html}{Quarto project} directory. A character scalar.}
}
\value{
A character vector.
}
\description{
Lists all \href{https://quarto.org/docs/projects/profiles.html}{Quarto project profiles} under the specified \code{input} path.
}
\examples{
if (quarto::is_using_quarto()) {
  quappo::profiles()
}
}
\seealso{
Other metadata: 
\code{\link{collapse_authors}()},
\code{\link{cur_profiles}()},
\code{\link{qmd_layout_classes}}
}
\concept{metadata}
