% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\docType{data}
\name{qmd_layout_classes}
\alias{qmd_layout_classes}
\title{Quarto article layout classes}
\format{
A character vector.
}
\usage{
qmd_layout_classes
}
\description{
All possible Quarto \href{https://quarto.org/docs/authoring/article-layout.html#available-columns}{article layout classes}.
}
\examples{
quappo::qmd_layout_classes
}
\seealso{
Other code chunk functions:
\code{\link{chunks}()},
\code{\link{fig_chunk}()},
\code{\link{read_chunk_toml}()},
\code{\link{tbl_chunk}()}
}
\concept{chunks}
\keyword{datasets}
