% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo-package.R
\docType{package}
\name{quappo-package}
\alias{quappo}
\alias{quappo-package}
\title{quappo: Unofficial Helper Functions around Quarto}
\description{
Provides unofficial helper functions around the quarto package, the R interface to the Quarto CLI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://quappo.rpkg.dev}
  \item \url{https://gitlab.com/rpkg.dev/quappo}
  \item Report bugs at \url{https://gitlab.com/rpkg.dev/quappo/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
