% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\name{read_chunk_toml}
\alias{read_chunk_toml}
\title{Read in Quarto code chunk data}
\usage{
read_chunk_toml(path)
}
\arguments{
\item{path}{Path to the TOML file containing structured figure or table code chunk data. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Reads in a \href{https://toml.io/en/latest}{TOML} file containing structured figure or table code chunk data, suited to be fed to \code{\link[=chunks]{chunks()}}.
}
\seealso{
Other code chunk functions:
\code{\link{chunks}()},
\code{\link{fig_chunk}()},
\code{\link{tbl_chunk}()}
}
\concept{chunks}
