% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\name{render_partly}
\alias{render_partly}
\title{Render Quarto book partly}
\usage{
render_partly(
  path = rstudioapi::getActiveDocumentContext()$path,
  profile = default_profiles(),
  output_format = NULL,
  execute = TRUE,
  execute_daemon_restart = FALSE,
  execute_debug = FALSE,
  use_freezer = FALSE,
  cache_refresh = FALSE,
  debug = FALSE,
  quiet = FALSE,
  pandoc_args = NULL,
  as_job = getOption("quarto.render_as_job", default = interactive() &&
    rstudioapi::isAvailable())
)
}
\arguments{
\item{path}{Path to a file (\href{https://quarto.org/docs/books/book-structure.html#overview}{"chapter"}) or subdirectory
(\href{https://quarto.org/docs/books/book-structure.html#parts-appendices}{"part"}) of a \href{https://quarto.org/docs/reference/projects/books.html}{Quarto book}
project.}

\item{profile}{\href{https://quarto.org/docs/projects/profiles.html}{Quarto project profile(s)} to use. Either
a character vector of profile names or \code{NULL} to use the default profile.}

\item{output_format}{Target output format (defaults to \code{"html"}). The option
\code{"all"} will render all formats defined within the file or project.}

\item{execute}{Whether to execute embedded code chunks.}

\item{execute_daemon_restart}{Restart keepalive Jupyter kernel before render.
Note this option is only applicable for rendering Jupyter notebooks or
Jupyter markdown.}

\item{execute_debug}{Show debug output for Jupyter kernel.}

\item{use_freezer}{Force use of frozen computations for an incremental
file render.}

\item{cache_refresh}{Force refresh of execution cache.}

\item{debug}{Leave intermediate files in place after render.}

\item{quiet}{Suppress warning and other messages.}

\item{pandoc_args}{Additional command line arguments to pass on to Pandoc.}

\item{as_job}{Render as an RStudio background job. Default is \code{"auto"},
which will render individual documents normally and projects as
background jobs. Use the \code{quarto.render_as_job} \R option to control
the default globally.}
}
\value{
\code{path}, invisibly.
}
\description{
Renders a file (\href{https://quarto.org/docs/books/book-structure.html#overview}{"chapter"}) or subdirectory
(\href{https://quarto.org/docs/books/book-structure.html#parts-appendices}{"part"}) of a \href{https://quarto.org/docs/reference/projects/books.html}{Quarto book}
project to the specified \code{profile}. Convenience wrapper around \code{\link[quarto:quarto_render]{quarto::quarto_render()}}, which by default renders the file that is \emph{currently active in
RStudio}.
}
\details{
This function is also registered as an \href{https://rstudio.github.io/rstudioaddins/}{RStudio add-in}, allowing RStudio users to assign a \href{https://support.rstudio.com/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts-in-the-RStudio-IDE}{custom shortcut} to it and to invoke it from the \href{https://www.rstudio.com/blog/rstudio-1-4-a-quick-tour/#command-palette--shortcuts}{command palette}.
}
\examples{
\dontrun{
# only works from inside a Quarto project
quappo::render_partly()}
}
