% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\name{view_output}
\alias{view_output}
\title{View Quarto output}
\usage{
view_output(
  input = ".",
  output_format = "html",
  profile = NULL,
  host = "127.0.0.1",
  port = 3456L,
  in_rstudio_viewer = interactive()
)
}
\arguments{
\item{input}{Path to \href{https://quarto.org/docs/projects/quarto-projects.html}{Quarto project} directory. A character scalar.}

\item{output_format}{Quarto output format to open. If \href{https://quarto.org/docs/reference/formats/html.html}{\code{"html"}}, a local HTTP server is launched
serving the projects's HTML output under \code{host} on the specified \code{port}.}

\item{profile}{\href{https://quarto.org/docs/projects/profiles.html}{Quarto project profile(s)} to use. Either
a character vector of profile names or \code{NULL} to use the default profile.}

\item{host}{IPv4 address to serve the HTML output on. A character scalar. Only relevant for \code{format = "html"}.}

\item{port}{TCP port number to serve the HTML output on. An integer scalar. Note that on most Unix-like systems including Linux and macOS, port numbers
below 1024 require root privileges. If the specified \code{port} is already occupied, a randomly chosen port between 1025–49151 is used instead. Only relevant
for \code{format = "html"}.}

\item{in_rstudio_viewer}{Whether to open the output file in the RStudio viewer pane if possible, i.e. RStudio \link[rstudioapi:isAvailable]{is available}. If
\code{FALSE} or RStudio is unavailable, the output file is opened in the system's default application for \code{output_format} via \code{\link[xopen:xopen]{xopen::xopen()}} instead.}
}
\value{
For \code{format = "html"}, a \code{\link[httpuv:WebServer]{httpuv::WebServer}} object. For \code{format = "pdf"}, the path to the PDF file. Invisibly in both cases.
}
\description{
Opens a Quarto project's pre-rendered output of the specified \code{format} in the RStudio viewer pane or the system's default application. For \code{"html"}, a local
HTTP server is launched that serves the HTML output on \href{http://127.0.0.1:3456}{\verb{127.0.0.1:3456}} by default.
}
\details{
Note that the web server started by \code{view_output()} will continue running in the background until it is explicitly stopped (e.g. via \code{\link[httpuv:stopServer]{httpuv::stopServer()}}
or \code{\link[httpuv:stopAllServers]{httpuv::stopAllServers()}}) or the \R session is terminated. An already running web server on \code{port} is reused if possible (i.e. if it was launched in the
same \R session), otherwise a port within the range 1024–49151 (\href{https://en.wikipedia.org/wiki/Registered_port}{"user" or "registered" ports}) is randomly
chosen using \code{\link[httpuv:randomPort]{httpuv::randomPort()}}.
}
\examples{
\dontrun{
# start server
s <- quappo::view_output()

# stop server
s$stop()}
}
