% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quappo.gen.R
\docType{data}
\name{qmd_layout_classes}
\alias{qmd_layout_classes}
\title{Quarto article layout classes}
\format{
A character vector.
}
\usage{
qmd_layout_classes
}
\description{
All possible Quarto \href{https://quarto.org/docs/authoring/article-layout.html#available-columns}{article layout classes}.
}
\examples{
quappo::qmd_layout_classes
}
\seealso{
Other metadata: 
\code{\link{collapse_authors}()},
\code{\link{cur_profiles}()},
\code{\link{default_profiles}()},
\code{\link{profiles}()}
}
\concept{metadata}
\keyword{datasets}
